Move graphql code to api folder
