+++ src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './apolloClient';
+export * from './graphql.generated';
